requirements.txt
asgiref==3.11.0
Django==5.1
django-cors-headers==4.9.0
djangorestframework==3.16.1
sqlparse==0.5.4
typing_extensions==4.15.0
